added intents and comments on places we can improve
