+++ src/pages/Esdt/models/.keep
@@ -1,0 +1,1 @@
+instead of helpers folder use this to define types or models.
